Adding Multiple Users
Created database table for users, modified projects table to have an
associated user. Added form on default page to register a new user,
added PHP and JS validation to this form. Began adding result messages
to the footer after a form is processed. Additionally, changed the
JS animations for changing sections in the home view to slide left or
right, depending on current section.