View History (by project)
Added the functionality to view activities by project. Updated the
project list so that clicking "view" hides the current list of
activities and then builds a new one from an ajax call limiting the
activities to those included in the project. Also changed it so that
only projects with activities are in the view history list.