Activity List More/Less
Designed and implemented "more" and "less" controls to display more or
less activities. Added a new PHP object "Lister" that generates these
lists; working on removing this functionality from the models. Creates use case
for an interface in PHP, but have not implemented this interface.

Set up ajax calls to use the Lister object rather than the methods in
the model classes.